requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
numpy==1.11.0
Pillow==3.2.0
psycopg2==2.6.1
wheel==0.24.0
whitenoise==2.0.6
